+++ include/Jet.hpp
@@ -8,16 +8,16 @@ class Jet : public Particle {
     
 public:
     Jet();
-    Jet(double pt, double eta, double phi, double mass = 0, double puId = 0, double qgLikelihood = 0, double CSV = 0);
-    Jet(const TLorentzVector&, double puId = 0, double qgLikelihood = 0, double CSV = 0);
+    Jet(float pt, float eta, float phi, float mass = 0, float puId = 0, float qgLikelihood = 0, float CSV = 0);
+    Jet(const TLorentzVector&, float puId = 0, float qgLikelihood = 0, float CSV = 0);
     virtual ~Jet();
-    void set(double pt, double eta, double phi, double mass, double puId, double qgLikelihood, double CSV);
-    void set(double pt, double eta, double phi, double mass = 0);
+    void set(float pt, float eta, float phi, float mass, float puId, float qgLikelihood, float CSV);
+    void set(float pt, float eta, float phi, float mass = 0);
     void setCsv( float csv );
     float csv() const;
     
 private:
-    double puId, qgLikelihood, CSV; 
+    float puId, qgLikelihood, CSV; 
     
 };
