+++ include/Particle.hpp
@@ -12,20 +12,20 @@ class Particle {
     
 public:
     Particle();
-    Particle(double pt, double eta, double phi, double mass = 0);
+    Particle(float pt, float eta, float phi, float mass = 0);
     Particle(const TLorentzVector&);
     virtual ~Particle();
-    void set(double pt, double eta, double phi, double mass = 0);
+    void set(float pt, float eta, float phi, float mass = 0);
     const TLorentzVector& getLorentzVector() const;
-    double pt() const;
-    double eta() const;
-    double phi() const;
-    double mass() const;
-    double dR(const Particle&) const;
-    double dPhi(const Particle&) const;
-    double dEta(const Particle&) const;
-    double mpp(const Particle&) const;
-    double ptpp(const Particle&) const;
+    float pt() const;
+    float eta() const;
+    float phi() const;
+    float mass() const;
+    float dR(const Particle&) const;
+    float dPhi(const Particle&) const;
+    float dEta(const Particle&) const;
+    float mpp(const Particle&) const;
+    float ptpp(const Particle&) const;
     
 protected:
     TLorentzVector p4;
